crypt/blockchain/pyproject.toml
[tool.poetry]
name = "blockchain"
version = "0.1.0"
description = "Algorand smart contracts"
authors = ["Luke Williams <willil14@rpi.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
beaker-pyteal = "^1.0.0"
algokit-utils = "^1.0.1"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pip-audit = "*"
black = {extras = ["d"], version = "*"}
flake8 = "*"
mypy = "*"
pytest = "*"
pytest-cov = "*"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120




[tool.pytest.ini_options]
pythonpath = ["smart_contracts", "tests"]


[tool.mypy]
files = "smart_contracts/"
python_version = "3.10"
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
allow_untyped_defs = false
strict_equality = true

